Fix payment logic, NFC timing, and display bugs
change NFC scan timer to unsigned long to match millis();

move money-entry prompt out of keypad input and into pay

increase delays for clearer UX;

Correct mortgage/bankruptcy money transfers and arithmetic in pay() to prevent incorrect balances and properly transfer remaining funds.

+++ GameVault/NFCReader.h
@@ -69,7 +69,7 @@ Player* readCard() {
 }
 
 Player* scan() {
-  int initialTime = millis();
+  unsigned long initialTime = millis();
   Player *currentPlayer;
 
   do {
